+++ config/module/detect_file.go
@@ -13,8 +13,6 @@ func (d *FileDetector) Detect(src, pwd string) (string, bool, error) {
 		return "", false, nil
 	}
 
-	// Make sure we're using "/" even on Windows. URLs are "/"-based.
-	src = filepath.ToSlash(src)
 	if !filepath.IsAbs(src) {
 		if pwd == "" {
 			return "", true, fmt.Errorf(
@@ -23,11 +21,5 @@ func (d *FileDetector) Detect(src, pwd string) (string, bool, error) {
 
 		src = filepath.Join(pwd, src)
 	}
-
-	// Make sure that we don't start with "/" since we add that below
-	if src[0] == '/' {
-		src = src[1:]
-	}
-
-	return fmt.Sprintf("file:///%s", src), true, nil
+	return fmtFileURL(src), true, nil
 }