+++ helper/schema/field_writer_map.go
@@ -78,6 +78,8 @@ func (w *MapFieldWriter) set(addr []string, value interface{}) error {
 		fallthrough
 	case TypeInt:
 		fallthrough
+	case TypeFloat:
+		fallthrough
 	case TypeString:
 		return w.setPrimitive(addr, value, schema)
 	case TypeList:
@@ -235,8 +237,13 @@ func (w *MapFieldWriter) setPrimitive(
 		if err := mapstructure.Decode(v, &n); err != nil {
 			return fmt.Errorf("%s: %s", k, err)
 		}
-
 		set = strconv.FormatInt(int64(n), 10)
+	case TypeFloat:
+		var n float64
+		if err := mapstructure.Decode(v, &n); err != nil {
+			return fmt.Errorf("%s: %s", k, err)
+		}
+		set = strconv.FormatFloat(float64(n), 'G', -1, 64)
 	default:
 		return fmt.Errorf("Unknown type: %#v", schema.Type)
 	}