+++ terraform/context_test.go
@@ -4044,6 +4044,25 @@ func TestContextPlan_taint(t *testing.T) {
 	}
 }
 
+// Doing a Refresh (or any operation really, but Refresh usually
+// happens first) with a config with an unknown provider should result in
+// an error. The key bug this found was that this wasn't happening if
+// Providers was _empty_.
+func TestContextRefresh_unknownProvider(t *testing.T) {
+	m := testModule(t, "refresh-unknown-provider")
+	p := testProvider("aws")
+	p.ApplyFn = testApplyFn
+	p.DiffFn = testDiffFn
+	ctx := testContext(t, &ContextOpts{
+		Module:    m,
+		Providers: map[string]ResourceProviderFactory{},
+	})
+
+	if _, err := ctx.Refresh(); err == nil {
+		t.Fatal("should error")
+	}
+}
+
 func TestContextPlan_multiple_taint(t *testing.T) {
 	m := testModule(t, "plan-taint")
 	p := testProvider("aws")